requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.5
comtypes==1.1.14
et-xmlfile==1.1.0
idna==3.4
numpy==1.23.3
openai==0.23.0
openpyxl==3.0.10
pandas==1.4.4
pandas-stubs==1.4.4.220912
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.21.0
pyttsx3==2.90
pytz==2022.2.1
pywin32==304
requests==2.28.1
six==1.16.0
speech==0.5.2
SpeechRecognition==3.8.1
tqdm==4.64.1
types-pytz==2022.2.1.0
typing_extensions==4.3.0
urllib3==1.26.12
